test: break out helpers_test
